error calculation in stump assessor corrected, appropriate unit tests fixed

+++ lib/src/optimizer/non_linear/decision_tree/assessor/stump_assessor.dart
@@ -1,6 +1,6 @@
 import 'package:ml_linalg/matrix.dart';
 
 abstract class StumpAssessor {
-  double getErrorOnStump(Iterable<Matrix> observations);
-  double getErrorOnNode(Matrix observations);
+  int getErrorOnStump(Iterable<Matrix> observations);
+  int getErrorOnNode(Matrix observations);
 }
